Support 'Skip to main content' button
# What
Upstream rejected the suggested changes for 'Skip to main content':
- https://github.com/cotes2020/jekyll-theme-chirpy/pull/1613

I've made the changes locally for this blog, and written `bin/diff-upstream-gem` as a simple helper script to keep an eye on any changes upstream that I might need to merge in.
To make it as simple as possible, I've added comments to the few lines that I've had to modify to make this work, so I can see, at a glance, when the different lines are my changes, versus the maintainers changes.